requirements.txt
mkdocs
mkdocs-material
click
jinja2
python-slugify
pyyaml